Extrai fluxo de cadastro em novo arquivo e widget - SignInPage
