+++ poc_amplitude/amplitude_poc/lib/pages/home_page.dart
@@ -1,9 +1,8 @@
-import 'package:amplitude_poc/app_state.dart';
-import 'package:amplitude_poc/pages/page01.dart';
 import 'package:amplitude_poc/pages/settings_page.dart';
-import 'package:amplitude_poc/resources/strings.dart';
 import 'package:amplitude_poc/widgets/nav_bar.dart';
 import 'package:flutter/material.dart';
+
+import 'sign_in_page.dart';
 
 class HomePage extends StatefulWidget {
   @override
@@ -54,46 +53,6 @@ class _HomePageState extends State<HomePage> {
           NavBarItem(iconData: Icons.timer),
           NavBarItem(iconData: Icons.settings),
         ],
-      ),
-    );
-  }
-}
-
-class SignInPage extends StatefulWidget {
-  const SignInPage({
-    Key key,
-  }) : super(key: key);
-
-  @override
-  _SignInPageState createState() => _SignInPageState();
-}
-
-class _SignInPageState extends State<SignInPage> {
-  void _startFlow() {
-    AppState.of(context).logEvent(
-      EventType.BUTTON_PRESSED,
-      'HomePage._onPressed',
-    );
-
-    Navigator.push(context, MaterialPageRoute(builder: (_) {
-      return PageOne();
-    }));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: RaisedButton.icon(
-        onPressed: _startFlow,
-        icon: Icon(Icons.adb_rounded, color: Colors.white),
-        color: Colors.pink,
-        label: Text(
-          Strings.START_FLOW,
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 16.0,
-          ),
-        ),
       ),
     );
   }
